move scene folder into component folder
